Fixed postponed matches
The premier league postponed two matches in gameweek 28 due to the FA Cup.
To do this, they have the gameweeks as null for the matches. This originally
caused a bug in my program so I now first check to see if the matchweek
is null before I add them to the fixture list.